# cars/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple CRUD API for profiles",
        "title": "Profiles API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4200",
    "basePath": "/",
    "paths": {
        "/admin/cars/delete/{id}": {
            "delete": {
                "description": "Delete all offers by user email",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete all offers",
                "operationId": "delete-offer-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the offer to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    }
                }
            }
        },
...
        "/cars/delete/all/{email}": {
            "delete": {
                "description": "Delete all offers by user email",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete all offers",
                "operationId": "delete-all-offers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address of the user to delete all offers",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    }
                }
            }
        },
        "/cars/delete/{email}": {
            "delete": {
                "description": "Delete all offers by user email",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete offer",
                "operationId": "delete-offer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address of the user to delete offer",
                        "name": "email",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ID of the offer to delete",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    }
                }
            }
        },
...
        "/cars/edit/{email}": {
            "put": {
                "description": "Edit a user offer by email",
                "produces": [
                    "application/json"
                ],
                "summary": "Edit an offer",
                "operationId": "edit-offer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address of the offer to be edited",
                        "name": "email",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User data to be edited",
                        "name": "userData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EditOffer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.EditOffer": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 3000
                },
                "id": {
                    "type": "string"
                },
                "mileage": {
                    "type": "integer"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "models.Id": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "responses.UserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}